+++ CHANGES.md
@@ -1,10 +1,10 @@
 Libplanet changelog
 ===================
 
-Version 2.6.0
+Version 3.0.0
 -------------
 
-To be released.
+Released on July 19, 2023.
 
 ### Backward-incompatible API changes